+++ R/fc_filter.R
@@ -21,6 +21,8 @@
 #' @param text_padding Changes the text padding inside the box. Default is 1. This number has to be greater than 0.
 #' @param bg_fill Box background color. It is white by default. See the `fill` parameter for \code{\link{gpar}}.
 #' @param border_color Box border color. It is black by default. See the `col` parameter for \code{\link{gpar}}.
+#' @param width Width of the box. If NA, it automatically adjusts to the content (default). Must be an object of class \code{\link{unit}} or a number between 0 and 1.
+#' @param height Height of the box. If NA, it automatically adjusts to the content (default). Must be an object of class \code{\link{unit}} or a number between 0 and 1.
 #' @param just_exc Justification for the text of the exclude box: left, center or right. Default is center.
 #' @param text_color_exc Color of the text of the exclude box. It is black by default. See `text_color`.
 #' @param text_fs_exc Font size of the text of the exclude box. It is 6 by default. See `text_fs`.
@@ -30,6 +32,8 @@
 #' @param bg_fill_exc Exclude box background color. It is white by default. See `bg_fill`.
 #' @param border_color_exc Box background color of the exclude box. It is black by default. See `border_color`.
 #' @param offset_exc Amount of space to add to the distance between the box and the excluded box (in the x coordinate). If positive, this distance will be larger. If negative, it will be smaller. This number has to be at least between 0 and 1 (plot limits) and the resulting x coordinate cannot exceed these plot limits. The default is NULL (no offset).
+#' @param width_exc Width of the exclude box. If NA, it automatically adjusts to the content (default). Must be an object of class \code{\link{unit}} or a number between 0 and 1.
+#' @param height_exc Height of the box. If NA, it automatically adjusts to the content (default). Must be an object of class \code{\link{unit}} or a number between 0 and 1.
 #' @return List with the filtered dataset and the flowchart parameters with the resulting filtered box.
 #'
 #' @examples
@@ -40,7 +44,7 @@
 #'
 #' @export
 
-fc_filter <- function(object, filter = NULL, N = NULL, label = NULL, text_pattern = "{label}\n {n} ({perc}%)", perc_total = FALSE, show_exc = FALSE, direction_exc = "right", label_exc = "Excluded", text_pattern_exc = "{label}\n {n} ({perc}%)", sel_group = NULL, round_digits = 2, just = "center", text_color = "black", text_fs = 8, text_fface = 1, text_ffamily = NA, text_padding = 1, bg_fill = "white", border_color = "black", just_exc = "center", text_color_exc = "black", text_fs_exc = 6, text_fface_exc = 1, text_ffamily_exc = NA, text_padding_exc = 1, bg_fill_exc = "white", border_color_exc = "black", offset_exc = NULL) {
+fc_filter <- function(object, filter = NULL, N = NULL, label = NULL, text_pattern = "{label}\n {n} ({perc}%)", perc_total = FALSE, show_exc = FALSE, direction_exc = "right", label_exc = "Excluded", text_pattern_exc = "{label}\n {n} ({perc}%)", sel_group = NULL, round_digits = 2, just = "center", text_color = "black", text_fs = 8, text_fface = 1, text_ffamily = NA, text_padding = 1, bg_fill = "white", border_color = "black", width = NA, height = NA, just_exc = "center", text_color_exc = "black", text_fs_exc = 6, text_fface_exc = 1, text_ffamily_exc = NA, text_padding_exc = 1, bg_fill_exc = "white", border_color_exc = "black", offset_exc = NULL, width_exc = NA, height_exc = NA) {
 
   is_class(object, "fc")
   UseMethod("fc_filter")
@@ -51,15 +55,15 @@ fc_filter <- function(object, filter = NULL, N = NULL, label = NULL, text_patter
 #' @importFrom rlang .data
 #' @importFrom rlang :=
 
-fc_filter.fc <- function(object, filter = NULL, N = NULL, label = NULL, text_pattern = "{label}\n {n} ({perc}%)", perc_total = FALSE, show_exc = FALSE, direction_exc = "right", label_exc = "Excluded", text_pattern_exc = "{label}\n {n} ({perc}%)", sel_group = NULL, round_digits = 2, just = "center", text_color = "black", text_fs = 8, text_fface = 1, text_ffamily = NA, text_padding = 1, bg_fill = "white", border_color = "black", just_exc = "center", text_color_exc = "black", text_fs_exc = 6, text_fface_exc = 1, text_ffamily_exc = NA, text_padding_exc = 1, bg_fill_exc = "white", border_color_exc = "black", offset_exc = NULL) {
+fc_filter.fc <- function(object, filter = NULL, N = NULL, label = NULL, text_pattern = "{label}\n {n} ({perc}%)", perc_total = FALSE, show_exc = FALSE, direction_exc = "right", label_exc = "Excluded", text_pattern_exc = "{label}\n {n} ({perc}%)", sel_group = NULL, round_digits = 2, just = "center", text_color = "black", text_fs = 8, text_fface = 1, text_ffamily = NA, text_padding = 1, bg_fill = "white", border_color = "black", width = NA, height = NA, just_exc = "center", text_color_exc = "black", text_fs_exc = 6, text_fface_exc = 1, text_ffamily_exc = NA, text_padding_exc = 1, bg_fill_exc = "white", border_color_exc = "black", offset_exc = NULL, width_exc = NA, height_exc = NA) {
 
   filter <- paste(deparse(substitute(filter)), collapse = "")
   filter <- gsub("  ", "", filter)
 
   if(filter == "NULL" & is.null(N)) {
-    stop("Either `filter` or `N` arguments have to be specified.")
+    cli::cli_abort("Either {.arg filter} or {.arg N} arguments must be specified.")
   }else if(filter != "NULL" & !is.null(N)) {
-    stop("`filter` and `N` arguments cannot be specified simultaneously.")
+    cli::cli_abort("The {.arg filter} and {.arg N} arguments cannot be specified simultaneously.")
   }
 
   if(!is.null(sel_group)) {
@@ -71,12 +75,17 @@ fc_filter.fc <- function(object, filter = NULL, N = NULL, label = NULL, text_pat
       }
 
       if(!any(sel_group %in% object$fc$group)) {
-        stop(stringr::str_glue("The specified `sel_group` does not match any group of the flowchart. Found groups in the flowchart are:\n{paste(object$fc$group[!is.na(object$fc$group)], collapse = '\n')}"))
+        cli::cli_abort(
+          c(
+            "The specified {.arg sel_group} does not match any group of the flowchart.",
+            "i" = "Found groups in the flowchart: {object$fc$group[!is.na(object$fc$group)]}"
+          )
+        )
       }
 
     } else {
 
-      stop("The `sel_group' argument cannot be given because no groups are found in the flowchart, as no previous split has been performed.")
+      cli::cli_abort("Cannot supply {.arg sel_group} because no groups exist in the flowchart yet, as no previous split has been performed.")
 
     }
 
@@ -88,15 +97,15 @@ fc_filter.fc <- function(object, filter = NULL, N = NULL, label = NULL, text_pat
 
     if(is.null(attr(object$data, "groups"))) {
       if(length(N) > 1) {
-        stop("The length of `N` has to be 1.")
+        cli::cli_abort("The length of {.arg N} has to be 1.")
       }
     } else {
       if(length(N) != nrow(attr(object$data, "groups"))) {
         if(is.null(sel_group)) {
-          stop(stringr::str_glue("The length of `N` has to match the number of groups in the dataset: {nrow(attr(object$data, 'groups'))}"))
+          cli::cli_abort("The length of {.arg N} has to match the number of groups in the dataset: {nrow(attr(object$data, 'groups'))}")
         } else {
           if(length(N) != length(sel_group)) {
-            stop(stringr::str_glue("The length of `N` has to match the number of selected groups in `sel_group`"))
+            cli::cli_abort("The length of {.arg N} has to match the number of selected groups in {.arg sel_group}")
           }
         }
       }
@@ -121,7 +130,7 @@ fc_filter.fc <- function(object, filter = NULL, N = NULL, label = NULL, text_pat
 
         } else {
 
-          stop(stringr::str_glue("The specified `sel_group` is not a grouping variable of the data. It has to be one of: {paste(tbl_groups$groups, collapse = ' ')}"))
+          cli::cli_abort("The specified {.arg sel_group} is not a grouping variable of the data. It has to be one of: {tbl_groups$groups}")
 
         }
 
@@ -129,7 +138,7 @@ fc_filter.fc <- function(object, filter = NULL, N = NULL, label = NULL, text_pat
 
       filt_rows <- unlist(purrr::map(1:nrow(tbl_groups), function (x) {
         if(N[x] > length(tbl_groups$.rows[[x]])) {
-          stop("The number of rows after the filter specified in N can't be greater than the original number of rows")
+          cli::cli_abort("The number of rows after the filter specified in {.arg N} can't be greater than the original number of rows.")
         } else {
           tbl_groups$.rows[[x]][1:N[x]]
         }
@@ -140,7 +149,7 @@ fc_filter.fc <- function(object, filter = NULL, N = NULL, label = NULL, text_pat
       nrows <- 1:nrow(object$data)
 
       if(N > length(nrows)) {
-        stop("The number of rows after the filter specified in N can't be greater than the original number of rows")
+        cli::cli_abort("The number of rows after the filter specified in {.arg N} cannot exceed the original number of rows.")
       }
 
       filt_rows <- nrows[1:N]
@@ -175,22 +184,34 @@ fc_filter.fc <- function(object, filter = NULL, N = NULL, label = NULL, text_pat
     )
 
   if(is.null(group0)) {
+
     new_fc$group <- NA
-  } else {
+
     new_fc <- new_fc |>
-      tidyr::unite("group", c(tidyselect::all_of(group0)), sep = " // ", na.rm = TRUE)
+      dplyr::left_join(object$fc |> dplyr::filter(.data$type != "exclude") |> dplyr::select("x", "group"), by = "group") |>
+      dplyr::group_by(.data$group) |>
+      dplyr::slice_tail(n = 1) |>
+      dplyr::ungroup()
+
+  } else {
+
+    new_fc <- new_fc |>
+      dplyr::mutate_at(tidyselect::all_of(group0), ~dplyr::case_when(is.na(.) ~ "NA", .default = .))|>
+      tidyr::unite("group", c(tidyselect::all_of(group0)), sep = " // ", na.rm = TRUE) |>
+      dplyr::left_join(object$fc |> dplyr::filter(.data$type != "exclude") |> dplyr::select("x", "group"), by = "group") |>
+      dplyr::mutate(group = factor(.data$group, levels = unique(.data$group))) |>
+      dplyr::group_by(.data$group) |>
+      dplyr::slice_tail(n = 1) |>
+      dplyr::ungroup() |>
+      dplyr::mutate(group = as.character(.data$group))
+
   }
 
-  new_fc <- new_fc |>
-    dplyr::left_join(object$fc |> dplyr::filter(.data$type != "exclude") |> dplyr::select("x", "group"), by = "group") |>
-    dplyr::group_by(.data$group) |>
-    dplyr::slice_tail(n = 1) |>
-    dplyr::ungroup()
 
   if(perc_total) {
     N_total <- unique(
       object$fc |>
-        dplyr::filter(is.na(.data$group)) |>
+        dplyr::filter(.data$y == max(.data$y)) |>
         dplyr::pull("N")
     )
     new_fc <- new_fc |>
@@ -205,7 +226,7 @@ fc_filter.fc <- function(object, filter = NULL, N = NULL, label = NULL, text_pat
   }
 
   if(text_padding == 0 | text_padding_exc == 0) {
-    stop("Text padding cannot be equal to zero")
+    cli::cli_abort("Text padding cannot be equal to zero.")
   }
 
   new_fc <- new_fc |>
@@ -220,7 +241,9 @@ fc_filter.fc <- function(object, filter = NULL, N = NULL, label = NULL, text_pat
       text_ffamily = text_ffamily,
       text_padding = text_padding,
       bg_fill = bg_fill,
-      border_color = border_color
+      border_color = border_color,
+      width = width,
+      height = height
     ) |>
     dplyr::ungroup() |>
     dplyr::select(-N_total)
@@ -241,7 +264,7 @@ fc_filter.fc <- function(object, filter = NULL, N = NULL, label = NULL, text_pat
 
     } else {
 
-      stop("The label has to be either a character or an expression.")
+      cli::cli_abort("The {.arg label} must be a character or an expression.")
 
     }
 
@@ -254,13 +277,13 @@ fc_filter.fc <- function(object, filter = NULL, N = NULL, label = NULL, text_pat
   if(is.null(sel_group)) {
 
     new_fc <- new_fc |>
-      dplyr::select("x", "y", "n", "N", "perc", "text", "type", "group", "just", "text_color", "text_fs", "text_fface", "text_ffamily", "text_padding", "bg_fill", "border_color")
+      dplyr::select("x", "y", "n", "N", "perc", "text", "type", "group", "just", "text_color", "text_fs", "text_fface", "text_ffamily", "text_padding", "bg_fill", "border_color", "width", "height")
 
   } else {
 
     new_fc <- new_fc |>
       dplyr::filter(.data$group %in% sel_group) |>
-      dplyr::select("x", "y", "n", "N", "perc", "text", "type", "group", "just", "text_color", "text_fs", "text_fface", "text_ffamily", "text_padding", "bg_fill", "border_color")
+      dplyr::select("x", "y", "n", "N", "perc", "text", "type", "group", "just", "text_color", "text_fs", "text_fface", "text_ffamily", "text_padding", "bg_fill", "border_color", "width", "height")
 
   }
 
@@ -335,14 +358,19 @@ fc_filter.fc <- function(object, filter = NULL, N = NULL, label = NULL, text_pat
 
     if(!is.null(offset_exc)) {
       if(!all(new_fc2$x >= 0 & new_fc2$x <= 1)) {
-        stop("The x-coordinate cannot exceed the plot limits 0 and 1. The argument offset_exc has to be set to a smaller number.")
+        cli::cli_abort(
+          c(
+            "The x-coordinate cannot exceed the plot limits 0 and 1.",
+            "i" = "The argument {.arg offset_exc} has to be set to a smaller number."
+          )
+        )
       }
     }
 
     if(perc_total) {
       N_total <- unique(
         object$fc |>
-          dplyr::filter(is.na(.data$group)) |>
+          dplyr::filter(.data$y == max(.data$y)) |>
           dplyr::pull("N")
       )
       new_fc2 <- new_fc2 |>
@@ -368,7 +396,9 @@ fc_filter.fc <- function(object, filter = NULL, N = NULL, label = NULL, text_pat
         text_ffamily = text_ffamily_exc,
         text_padding = text_padding_exc,
         bg_fill = bg_fill_exc,
-        border_color = border_color_exc
+        border_color = border_color_exc,
+        width = width_exc,
+        height = height_exc
       ) |>
       dplyr::select(-"parent", -"N_total")
 
@@ -388,7 +418,7 @@ fc_filter.fc <- function(object, filter = NULL, N = NULL, label = NULL, text_pat
 
       } else {
 
-        stop("The label_exc has to be either a character or an expression.")
+        cli::cli_abort("The {.arg label_exc} has to be either a character or an expression.")
 
       }
 
@@ -426,7 +456,6 @@ fc_filter.fc <- function(object, filter = NULL, N = NULL, label = NULL, text_pat
 
   } else {
 
-    #in development
     groups <- names(attr(object$data, "groups"))
     groups <- groups[groups != ".rows"]